MNT: separate find-replace utilities from widgets
